VueDRY: Move FontSize directive name from constants to it's declaration, add constant with default FontStyle, update its usages
