+++ src/constants/styles.ts
@@ -1,2 +1,4 @@
-export const FontSizeDirectiveName: string = 'font-size';
-export const FontSizeDefaultValue: number = 14;
+import { FontStyle } from '@/enums/styles';
+
+export const DefaultFontSize: number = 14;
+export const DefaultFontStyle: FontStyle = FontStyle.Normal;
