+++ src/mixins/buttonMixin.ts
@@ -1,21 +1,21 @@
-import FontDirective from '@/directives/fontSizeDirective';
 import Vue from 'vue';
-import { FontSizeDefaultValue, FontSizeDirectiveName } from '@/constants/styles';
+import { DefaultFontSize, DefaultFontStyle } from '@/constants/styles';
+import { FontSizeDirective, FontSizeDirectiveName } from '@/directives/fontSizeDirective';
 import { FontStyle } from '@/enums/styles';
 import { RecordPropsDefinition } from 'vue/types/options';
 
 export default {
     directives: {
-        [FontSizeDirectiveName]: FontDirective,
+        [FontSizeDirectiveName]: FontSizeDirective,
     },
     props: {
         fontSize: {
             type: Number,
-            default: FontSizeDefaultValue,
+            default: DefaultFontSize,
         },
         fontStyle: {
             type: String,
-            default: FontStyle.Normal,
+            default: DefaultFontStyle,
         },
     } as RecordPropsDefinition<{ fontSize: number; fontStyle: FontStyle }>,
     methods: {
